Add files via upload
This program is choosing a randomly selected movie for the user to watch from a JSON database, which contains two types of movies, light-hearted comedies or dark, heavy movies. The user can choose from these categories, update the database with new movies. After a title was presented by the program, the user can decide if he/she wants to delete it from the database.